refactor: Extension 관련 리팩토링
